Add feature to check previous readings
Feature allows the last valid readings to be stored, and in turn read
the next time the application is run. This is then used in conjunction
with the evaluate method to check the consistency.

This has been added to confront the issue found where it appears that
some bits are being shifted in such a way that the values are not
correct, but the checksum is still acceptable. This is often expressed
as doubled humidity and temperature values, which can be identified
by the humidity going above the 100% value, but this is not always
the case.

If the last stored values are inconsistent, the current reading will be
stored, then once another reading is obtained, this will be compared
firstly to the stored value, and then to the previously read value.

Assumptions made:
* Temperature and Humidity values should not change more than 5*c or 5%
between the current value and the last read value.
* Readings are made every minute.
* If two values read during the current evaluation both disagree with
the stored value, assume the readings are now correct and they have
changed significantly since the last stored value.

Change-Id: I3d7bfa37f61230743a439bdbad2c4b265c444c56